+++ CHANGELOG.rst
@@ -2,6 +2,35 @@
 Changelog
 =========
 
+0.5.6 (2019-02-21)
+==================
+
+Fixes
+-----
+* Set row.title to "" instead of None to get row behavior.
+
+
+0.5.5 (2019-02-13)
+==================
+
+Fixes
+-----
+* Merged in https://github.com/weaveworks/grafanalib/pull/143 to support custom variables
+
+
+0.5.4 (2019-02-08)
+==================
+
+Fixes
+-----
+* Added more base units
+
+
+New features
+------------
+* Added an AJAX panel plugin object
+
+
 0.5.2 (2018-07-19)
 ==================